backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
vapi_server_sdk
supabase
pydantic-settings
requests
google-generativeai==0.8.3
Pillow 